test/pyproject.toml
[project]
name = "ai-hackathon-2025-test"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "RAMEN team" }
]
dependencies = [
    "pytest-playwright>=0.5.1",
    "pytest-xdist>=3.6.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/template_qa_python"]
